Add files via upload
Fixed Project and Gerber. I found on the previous version a problem with the LIMIT resistor from TPS2553, wich was shorted. When I placed it on the project, the software did not cut the line and it was kept, making the resistor useless. Thos short was removed and the resistor (that limits the current to sensors VIN) is now functional.